[lib/NDLCom] on the way to a general NDLCom-Widget
This is a very cool Widget!

It provides the NDLCom::NDLComContainer-class, which can be easily added
to your application! Just do like in test/testGUI.cpp
{{{
    /* carefull! in ctor of this class, QDockWidgets will be inserted into this QMainWindow! */
    NDLCom::NDLComContainer* ndlcom = new NDLCom::NDLComContainer(this);
    mpUi->menubar->addMenu(ndlcom->mpMenu);
    addToolBar(ndlcom->mpToolbar);
}}}
and this widget will add its sub-widgets into your QDockWidget area. By
adding the menu- and toolbar into your application, all functionality is
made accessible to the user. Nice!

In the container-widget-class, new'ing of the needed sub-modules is
done. Connecting all internal signals and creating the QMenu and
QToolBar.

The following sub-widgets are provided:
- Composer() -- the second generation TelegramComposer. Better than ever!
- CommunicationStatistics -- real-time receiving rate of messages
- RepresentationsInfo -- statically showing all known representation-ids
- MessageTraffic -- showing which messages are globally going in an out.
- NDLCom -- of course, the interface to manage all individual connections.
  Allows showing raw-data-traffic of each interface.

This whole class is based on the NDLCom::Message class. This class does
a timestamp when its created (during receive normally) and helps in
tracking the order of packages from different interfaces. It also
contains the ProtocolHeader of this message as well as the decoded
raw-data as a char-array. This data can be used by other applications.
Or it can be casted to a Representations::<blabla> class, which is done
in NDLCom::RepresentationsMapper.

The NDLCom::NDLComContainer emits signals of type "const Message&"
(originating from NDLCom::NDLCom) as well as all currently known
Representations, correctly casted to the right one (originating from
RepresentationsMapper). It receives signals of "const Message&", which
are then encoded and transferred to all running interfaces.

Each connection is represented by the same base-class
NDLCom::Interface, which handles all the "low-level" routine work. For
example calling connect/dsconnect, pause/resume, setting the correct
icons, showing data-rate and raw-traffic. In the moment, two kinds of
real interfaces are provided -- serialcom and udpcom. Of both
sub-types, arbitrary number of connections can be opened inside
NDLCom::NDLCom, which then all are capable of sending and receiving
data. All connections can be closed on one click as well as individually,
via QToolBar or QMenu.

It was tried to use QAction as often as possible, to have all the actions in a
unified interface. This allows easy use of QToolBar and QMenu.

TODO:
- since we don't have a routing algorithm, all telegrams to send are
  emitted on all interfaces! And telegrams not directed to the PC are not
  forwarded. There is a forward-stub in the code which is deactivated...
- The raw-traffic of each interface is shown in a separate window. This is
  not nice and not in the way it was handled before. But I didn't run into
  a nicer solution, and it works this way.

FIXME:
- something with the NDLCom::Message() class is broken on accessing
  references... grep the code...
- Interface-raw-traffic as separate pop-up-window is not properly working
  in the moment, if one window is kept open, the GUI may not close properly...

Signed-off-by: Martin Zenzes (RH5) <martin.zenzes@dfki.de>

git-svn-id: https://svn.hb.dfki.de/iStruct/software/trunk@1970 23edfee6-d8fb-8640-80ab-3a9191bda3f1