[lib/NDLCom] don't delete interfaces too early
the removal from the runningInterfaces list is done later, when the
disconnected() signal from the interface arrives. so deleting the interface
from the list so early (by calling takeFirst()) may confuse others.

as far as i could see this did not cause any problems. but it makes the code
harder to read

additionally change on line from "emit" to the usage of
"activate(QAction::Trigger)" which should be more qt'ish

Signed-off-by: Martin Zenzes (RH5) <martin.zenzes@dfki.de>

git-svn-id: https://svn.hb.dfki.de/iStruct/software/trunk@3907 23edfee6-d8fb-8640-80ab-3a9191bda3f1