[lib/NDLCom/interface_container] catch cancel in connection dialog
i think this was a memory leak before: click "cancel" in the connection dialog
of an interface caused nobody to delete this thing, since the connect() signa
which triggered adding the interfaces pointer to the runningInterfaces list
never arrived. using a new interface property, we ask the interface if it is
working and delete it again if not.

Signed-off-by: Martin Zenzes (RH5) <martin.zenzes@dfki.de>

git-svn-id: https://svn.hb.dfki.de/iStruct/software/trunk@3909 23edfee6-d8fb-8640-80ab-3a9191bda3f1