[lib/NDLCom] add xml-solution for the deviceId-mess
after representations are not a safe place for deviceIds anymore, some
projects had to move to new ground. this is the result, and it has only
relevance in SeeGrip, note that! other projects may piggiback ;-)

generates:
- DeviceIds.c # gets linked into libNDLCom.a
- DeviceIds.h # resides in build/$ARCH/include/NDLCom, get it via pkg-config as usual
- Devices.vhd # gets into ${CMAKE_CURRENT_SOURCE_DIR}/vhdl

the define-format (they are now enums) has slightly
changed, for the C-case for example:

{{{
    hdr.mReceiverId = ndlcomDeviceIdCAM_TCM8230MD;
}}}

or in c++:
{{{
    using ndlcom::deviceId;

    hdr.mReceiverId = CAM_TCM8230MD;
}}}

Signed-off-by: Martin Zenzes (RH5) <martin.zenzes@dfki.de>

git-svn-id: https://svn.hb.dfki.de/iStruct/software/trunk@6380 23edfee6-d8fb-8640-80ab-3a9191bda3f1