[lib/NDLCom] make "numberOfCrcFails" in ProtocolParserState a int16_t
also, don't reset it in the udp-case for a new packet. does not make a
difference, since in udp we don't have a crc, although...

also remove the "numberOfEscapes" from the ProtocolParserState struct. it
is not used anywhere in istruct nor seegrip and did not work anyways.
before it contained the value of the "lastByteWasCRC" state -- either zero
or one...

Signed-off-by: Martin Zenzes (RH5) <martin.zenzes@dfki.de>

git-svn-id: https://svn.hb.dfki.de/iStruct/software/trunk@7273 23edfee6-d8fb-8640-80ab-3a9191bda3f1