ndlcomBridge: change semantics of decoded-message-handling
all decoded messages are handled in a central place, no matter where they are from

only messages from external can update the routing table with their origin!

this leads to the "feature" that internal handlers can send messages to other internal handlers... the "ndlcomBridgeProcessDecodedMessage" tries to prevent loops by comparing the "origin" of the message

the beauty of C: no change in the interface at all!

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>