Bridge/Node: split into "node" and "bridge"
rationale: a "bridge" does not have a device-id... but a "node" does. and a "bridge" can have multiple "nodes", hence receive messages from multiple receiverIds

move some code from the "bridge" to the "node", no "ownIdHandler" list and this "ONLY_OWN_ID" flag from the "interface" is not needed anymore.

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>