ndlcomBridge: add pty support
opens pseudo-terminal and creates symlink at given location. allows the
CommonGUI to directly connect to a "serial port" behind the symlink.
baudrate does not matter. call bridge:

for undecernable routing of one usb port to the pseuso terminal do:

    ./ndlcomBridge -u pty:///tmp/symlink -f 10000 -u serial:///dev/ttyUSB1

and open the pty with the gui:

    ./CommonGUI --serialport=/tmp/testii

handy! as the bridge itself has a routing table, this can be used to
really connect route messages between ports independently of the gui.

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>