fix: fatal error when usb device disconnects
done via checking for "POLLHUP" so that the bridge at least stops
working when something is not right.

kinda expensive... this is an extra syscall per serial oder fpga
interface in the hotpath of "ndlcomBridgeProcess()"..

also removes some older comments mentioning usage of poll together with
the pty-interface to check for unwanted events. is not relevant anymore,
and only confuses.

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>