Bridge: fix bug in internal message handling
internal handlers _could opt to not see messages coming from other internal agents. the boolean for this was bugged (missing parenthesis) so that the flag was turned around -- all handlers opted to not see internal messages. now they do.

note that the behavior of the bridge is now changed!

also some clarification on a section of comment.

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>