improve "printStatus" functionality
recurse. recurse all the things!

there is a problem, if an inheriting class again calls
"HandlerCommon::printStatus" and add its own context... how to ensure
consistent output when having a long inheritance tree, with additional
output of childclasses?

ideally i want to have the output the be a proper json datastructure in
the end... not there yet... next time

Signed-off-by: Martin Zenzes <martin.zenzes@dfki.de>